Remove unnecessary static_cast

+++ mqt/MqtBase/MatchContact.cpp
@@ -9,12 +9,14 @@
 #include "base_pch.h"
 
 //---------------------------------------------------------------------------
+
 QSharedPointer<MatchContact> BaseMatchContest::pcontactAt( int i )
 {
-    if (i > static_cast< int > (contactMatchList.size()))
+    if (i > contactMatchList.size())
         return QSharedPointer<MatchContact>();
     return std::next(contactMatchList.begin(), i)->wt;
 }
+
 // Here we compare pointers, just to give an actual < operator
 bool MatchContactList::operator<( const BaseMatchContest& rhs ) const
 {
